requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-ninja==0.12.2
idna==2.10
pkg-resources==0.0.0
pydantic==1.7.3
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
